+++ src/main/java/net/md_5/bungee/jni/zlib/NativeZlib.java
@@ -37,7 +37,7 @@ public class NativeZlib implements BungeeZlib
     }
 
     @Override
-    public long process(ByteBuf in, ByteBuf out) throws DataFormatException
+    public int process(ByteBuf in, ByteBuf out) throws DataFormatException
     {
         // Smoke tests
         in.memoryAddress();
@@ -54,7 +54,7 @@ public class NativeZlib implements BungeeZlib
             out.writerIndex( out.writerIndex() + processed );
         }
 
-        long adler = nativeCompress.checksum( ctx );
+        int adler = nativeCompress.checksum( ctx );
         nativeCompress.reset( ctx, compress );
         nativeCompress.consumed = 0;
         nativeCompress.finished = false;
